requirements.txt
requests>=2.25.0
python-dotenv>=0.15.0  # <-- added dependency to load .env file
streamlit>=1.0.0  # <-- added dependency for Streamlit
